OTLP part 1 (#76)
Refactors prom remote write inflow and outflow into an HTTP
base since 99% of the code will be shared. Also add the
OTLP protos and compile them. Most of the diff in this PR
is proto changes.

The only functional change is I moved the
append_tags_to_downstream_id config option inside
DownstreamIdSource so it can more easily be shared. This is
a breaking config change but it should be OK given
current users.

Signed-off-by: Matt Klein <mklein@bitdrift.io>